% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_time_human}
\alias{get_time_human}
\title{get_time_human}
\source{
https://github.com/daattali/shiny-server/tree/master/persistent-data-storage
}
\usage{
get_time_human()
}
\value{

}
\description{
get a formatted string of the timestamp (exclude colons as they are invalid
}
\examples{
get_time_human()
}
