% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_jsWrappers.R
\name{scrollIntoView}
\alias{scrollIntoView}
\title{scrollIntoView}
\source{
https://stackoverflow.com/questions/61757698/how-to-go-to-an-html-anchor-programatically-in-shiny
}
\usage{
scrollIntoView(id)
}
\arguments{
\item{id}{the html anchor id}
}
\value{

}
\description{
Wrapper for js function scrollIntoView.
Scroll window to the anchor id.
}
