% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_jsWrappers.R
\name{scrollIntoView}
\alias{scrollIntoView}
\title{scrollIntoView}
\usage{
scrollIntoView(id)
}
\arguments{
\item{id}{the html anchor id}
}
\value{

}
\description{
Wrapper for js function scrollIntoView.
Scroll window to the anchor id.
}
